Fix size overflows with integers

+++ src/fft/fft_grid_layout.h
@@ -49,7 +49,7 @@ typedef struct {
   int npts_global_gspace[3];
   double cutoff;
   // Number of local points in g-space (relevant with ray-distribution)
-  int npts_gs_local;
+  size_t npts_gs_local;
   bool ray_distribution;
   int (*ray_to_xy)[2];
   int *xy_to_ray;
@@ -95,7 +95,7 @@ typedef struct {
   // redistribution object
   fft_redistribution_t *redistribution;
   // Buffers for FFTs
-  size_t buffer_size;
+  int buffer_size;
 } fft_grid_layout;
 
 /*******************************************************************************
